Extensive refactor to pull most of the code out of __main__
chunks.py is bearing the brunt of this, but this will let there be more info when debugging hopefully, as well as making the code easier to follow
